+++ pkg/api/query/v1.go
@@ -76,6 +76,7 @@ const (
 	ShardInfoParam           = "shard_info"
 	LookbackDeltaParam       = "lookback_delta"
 	EngineParam              = "engine"
+	QueryExplainParam        = "explain"
 )
 
 type PromqlEngineType string
@@ -272,7 +273,8 @@ type queryData struct {
 	Result     parser.Value     `json:"result"`
 	Stats      stats.QueryStats `json:"stats,omitempty"`
 	// Additional Thanos Response field.
-	Warnings []error `json:"warnings,omitempty"`
+	QueryExplanation *engine.ExplainOutputNode `json:"explanation,omitempty"`
+	Warnings         []error                   `json:"warnings,omitempty"`
 }
 
 func (qapi *QueryAPI) parseEnableDedupParam(r *http.Request) (enableDeduplication bool, _ *api.ApiError) {
@@ -419,6 +421,27 @@ func (qapi *QueryAPI) parseShardInfo(r *http.Request) (*storepb.ShardInfo, *api.
 	return &info, nil
 }
 
+func (qapi *QueryAPI) parseQueryExplainParam(r *http.Request, query promql.Query) (*engine.ExplainOutputNode, *api.ApiError) {
+	var explanation *engine.ExplainOutputNode
+
+	if val := r.FormValue(QueryExplainParam); val != "" {
+		var err error
+		enableExplanation, err := strconv.ParseBool(val)
+		if err != nil {
+			return explanation, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", QueryExplainParam)}
+		}
+		if enableExplanation {
+			if eq, ok := query.(engine.ExplainableQuery); ok {
+				explanation = eq.Explain()
+			} else {
+				return explanation, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Errorf("Query not explainable")}
+			}
+		}
+	}
+
+	return explanation, nil
+}
+
 func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiError, func()) {
 	ts, err := parseTimeParam(r, "time", qapi.baseAPI.Now())
 	if err != nil {
@@ -509,6 +532,11 @@ func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiErro
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
 
+	explanation, apiErr := qapi.parseQueryExplainParam(r, qry)
+	if apiErr != nil {
+		return nil, nil, apiErr, func() {}
+	}
+
 	tracing.DoInSpan(ctx, "query_gate_ismyturn", func(ctx context.Context) {
 		err = qapi.gate.Start(ctx)
 	})
@@ -541,9 +569,10 @@ func (qapi *QueryAPI) query(r *http.Request) (interface{}, []error, *api.ApiErro
 		qs = stats.NewQueryStats(qry.Stats())
 	}
 	return &queryData{
-		ResultType: res.Value.Type(),
-		Result:     res.Value,
-		Stats:      qs,
+		ResultType:       res.Value.Type(),
+		Result:           res.Value,
+		Stats:            qs,
+		QueryExplanation: explanation,
 	}, res.Warnings, nil, qry.Close
 }
 
@@ -667,6 +696,11 @@ func (qapi *QueryAPI) queryRange(r *http.Request) (interface{}, []error, *api.Ap
 		return nil, nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}, func() {}
 	}
 
+	explanation, apiErr := qapi.parseQueryExplainParam(r, qry)
+	if apiErr != nil {
+		return nil, nil, apiErr, func() {}
+	}
+
 	tracing.DoInSpan(ctx, "query_gate_ismyturn", func(ctx context.Context) {
 		err = qapi.gate.Start(ctx)
 	})
@@ -697,9 +731,10 @@ func (qapi *QueryAPI) queryRange(r *http.Request) (interface{}, []error, *api.Ap
 		qs = stats.NewQueryStats(qry.Stats())
 	}
 	return &queryData{
-		ResultType: res.Value.Type(),
-		Result:     res.Value,
-		Stats:      qs,
+		ResultType:       res.Value.Type(),
+		Result:           res.Value,
+		Stats:            qs,
+		QueryExplanation: explanation,
 	}, res.Warnings, nil, qry.Close
 }